pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer - analyzes specified pages for SEO-friendliness"
authors = ["danlo <danlo12@mail.ru>"]
readme = "README.md"

packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
gunicorn = "^21.2.0"
flask = "^3.0.0"
psycopg2-binary = "^2.9.9"
python-dotenv =  "^1.0.1"
beautifulsoup4 = "^4.12.3"
validators = "^0.28.1"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
